% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizationR.R
\docType{package}
\name{optimizationR-package}
\alias{optimizationR}
\alias{optimizationR-package}
\title{Decision Optimization with R - Tutorial Codes}
\description{
Open any tutorial code from "Decision Optimization with R". Furthermore update the tutorial contents anytime without updating the package.
}
\references{
\url{http://www.optimization-r.com/}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.optimization-r.com/}
}

}
\author{
Ronald Hochreiter, \email{ron@hochreiter.net}
}
