% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{packages.cran}
\alias{packages.cran}
\title{List of packages required from CRAN}
\format{One character vector}
\usage{
data(packages)
}
\description{
List of packages required from CRAN
}
\keyword{datasets}
