% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update the package contents without updating the package}
\usage{
update()
}
\description{
\code{update} updates the current list of tutorials without requiring
to update the package.
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
